Initial project setup with backend and configs
Add initial IntelliJ IDEA project files, backend module configuration, application properties, compiled classes, and Maven build outputs for the pillpath-backend project. Includes Spring Boot setup, JPA, PostgreSQL configuration, and basic controllers, DTOs, entities, repositories, and services.

+++ src/main/java/com/leo/pillpathbackend/service/UserService.java
@@ -4,9 +4,9 @@ import com.leo.pillpathbackend.dto.UserDTO;
 
 public interface UserService {
     UserDTO createUser(UserDTO userDTO);
-    UserDTO getUserById(int id);
+    UserDTO getUserById(Long id);
     UserDTO updateUser(UserDTO userDTO);
-    void deleteUser(int id);
+    void deleteUser(Long id);
     UserDTO getUserByUsername(String username);
     UserDTO getUserByEmail(String email);
     boolean validateUser(String username, String password);
